Move files out of mediawiki/ directory
This distinction isn’t very helpful at the moment – it seems better to
have all the data access stuff in data-access (and MwMessagesRepository
in MessagesPlugin/).
